fees > taxes
Changed 'fees' to 'taxes' where appropriate to better match the definition of the word with its context.



Alternatively a mechanism could be implemented such as proposed by tinkyholloway:

Only fee payers are counted as 'citizens'. Only citizens can participate in goverment: stand for office, vote, veto vote, be on juries, hold the title 'Citizen' etc. it's basically a flat fee, poll tax.

Definitions:
Fee: a payment made to a professional person or to a professional or public body in exchange for advice or services.
Tax: a compulsory contribution to state revenue, levied by the government on workers' income and business profits or added to the cost of some goods, services, and transactions.